+++ .idea/gradle.xml
@@ -8,14 +8,14 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
-            <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/MyRuns1" />
             <option value="$PROJECT_DIR$/backend" />
           </set>
         </option>
         <option name="myModules">
           <set>
             <option value="$PROJECT_DIR$" />
-            <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/MyRuns1" />
             <option value="$PROJECT_DIR$/backend" />
           </set>
         </option>
