+++ .idea/libraries/play_services_6_5_87.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="play-services-6.5.87">
     <CLASSES>
-      <root url="file://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.google.android.gms/play-services/6.5.87/res" />
-      <root url="jar://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.google.android.gms/play-services/6.5.87/jars/classes.jar!/" />
+      <root url="jar://$PROJECT_DIR$/MyRuns1/build/intermediates/exploded-aar/com.google.android.gms/play-services/6.5.87/jars/classes.jar!/" />
+      <root url="file://$PROJECT_DIR$/MyRuns1/build/intermediates/exploded-aar/com.google.android.gms/play-services/6.5.87/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
