+++ .idea/modules.xml
@@ -3,7 +3,7 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/MyRuns1.iml" filepath="$PROJECT_DIR$/MyRuns1.iml" />
-      <module fileurl="file://$PROJECT_DIR$/app/MyRuns1-MyRuns1.iml" filepath="$PROJECT_DIR$/app/MyRuns1-MyRuns1.iml" />
+      <module fileurl="file://$PROJECT_DIR$/MyRuns1/MyRuns1-MyRuns1.iml" filepath="$PROJECT_DIR$/MyRuns1/MyRuns1-MyRuns1.iml" />
       <module fileurl="file://$PROJECT_DIR$/backend/backend.iml" filepath="$PROJECT_DIR$/backend/backend.iml" />
     </modules>
   </component>
