+++ MyRuns1/src/main/AndroidManifest.xml
