+++ MyRuns1/src/main/java/edu/dartmouth/cs/camera/DisplayEntryActivity.java
