+++ MyRuns1/src/main/java/edu/dartmouth/cs/camera/MainActivity.java
