+++ MyRuns1/src/main/java/edu/dartmouth/cs/camera/SettingsFragment.java
