+++ MyRuns1/src/main/java/edu/dartmouth/cs/camera/database/ExerciseEntry.java
