+++ MyRuns1/src/main/java/edu/dartmouth/cs/camera/database/ExerciseEntryAsyncTask.java
