+++ MyRuns1/src/main/res/layout/activity_display_entry.xml
