+++ MyRuns1/src/main/res/values/strings.xml
