+++ MyRuns1/src/main/res/values/styles.xml
