+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':backend'
+include ':MyRuns1', ':backend'
